+++ src/Command/AddUserCreateCommand.php
@@ -1,0 +1,61 @@
+<?php
+
+namespace App\Command;
+
+use App\Entity\User;
+use Doctrine\ORM\EntityManagerInterface;
+use Symfony\Component\Console\Attribute\AsCommand;
+use Symfony\Component\Console\Command\Command;
+use Symfony\Component\Console\Input\InputArgument;
+use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Console\Style\SymfonyStyle;
+use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
+
+#[AsCommand(
+    name: 'app:user:add',
+    description: 'Add user for administrative access',
+)]
+class AddUserCreateCommand extends Command
+{
+    public function __construct(
+        private readonly EntityManagerInterface $entityManager,
+        private readonly UserPasswordHasherInterface $passwordHasher,
+    ) {
+        parent::__construct();
+    }
+
+    protected function configure(): void
+    {
+        $this->addArgument('email', InputArgument::REQUIRED, 'E-mail-address for username')
+            ->addArgument('password', InputArgument::REQUIRED, 'Users login password');
+    }
+
+    /**
+     * @todo: needs way better input validation.
+     */
+    protected function execute(InputInterface $input, OutputInterface $output): int
+    {
+        $user = new User();
+        $io = new SymfonyStyle($input, $output);
+
+        $email = $input->getArgument('email');
+
+        if (!filter_var($email, FILTER_VALIDATE_EMAIL)) {
+            $io->error('Email not valid');
+
+            return Command::FAILURE;
+        } else {
+            $user->setEmail($email);
+        }
+
+        $plainPassword = $input->getArgument('password');
+        $hashedPassword = $this->passwordHasher->hashPassword($user, $plainPassword);
+        $user->setPassword($hashedPassword);
+
+        $this->entityManager->persist($user);
+        $this->entityManager->flush();
+
+        return Command::SUCCESS;
+    }
+}