+++ src/Controller/Admin/PointOfInterestController.php
@@ -1,0 +1,40 @@
+<?php
+
+namespace App\Controller\Admin;
+
+use App\Entity\PointOfInterest;
+use EasyCorp\Bundle\EasyAdminBundle\Config\Asset;
+use EasyCorp\Bundle\EasyAdminBundle\Controller\AbstractCrudController;
+use EasyCorp\Bundle\EasyAdminBundle\Field\DateField;
+use EasyCorp\Bundle\EasyAdminBundle\Field\IdField;
+use EasyCorp\Bundle\EasyAdminBundle\Field\ImageField;
+use EasyCorp\Bundle\EasyAdminBundle\Field\TextField;
+use EasyCorp\Bundle\EasyAdminBundle\Form\Type\FileUploadType;
+
+class PointOfInterestController extends AbstractCrudController
+{
+    public static function getEntityFqcn(): string
+    {
+        return PointOfInterest::class;
+    }
+
+    public function configureFields(string $pageName): iterable
+    {
+        return [
+            IdField::new('id')->hideOnForm(),
+            TextField::new('name')->setRequired(true)
+                ->setHelp('Name this'),
+            TextField::new('latitude')->setRequired(true)
+                ->setHelp('The latitude of the interest point'),
+            TextField::new('longitude')->setRequired(true)
+                ->setHelp('The longitude of the interest point'),
+            TextField::new('subtitles')->setRequired(true)
+                ->setHelp('A text version of the podcast, for people with hearing disabilities.'),
+            ImageField::new('image')->setRequired(true)->setUploadDir('/public/points-of-interest')->hideWhenUpdating(),
+            TextField::new('podcast')->setRequired(true)->setFormType(FileUploadType::class)->addJsFiles(Asset::fromEasyAdminAssetPackage('field-file-upload.js'))->hideWhenUpdating(),
+            IdField::new('id')->hideOnForm(),
+            DateField::new('createdAt')->hideOnForm(),
+            DateField::new('updatedAt')->hideOnForm(),
+        ];
+    }
+}